Moved ScaleOption in its own file
